--- src/main/java/woowacourse/auth/dto/TokenResponse.java
@@ -1,16 +1,0 @@
-package woowacourse.auth.dto;
-
-public class TokenResponse {
-    private String accessToken;
-
-    public TokenResponse() {
-    }
-
-    public TokenResponse(String accessToken) {
-        this.accessToken = accessToken;
-    }
-
-    public String getAccessToken() {
-        return accessToken;
-    }
-}
