+++ src/main/java/woowacourse/shoppingcart/service/CustomerService.java
@@ -6,9 +6,9 @@ import woowacourse.auth.support.JwtTokenProvider;
 import woowacourse.shoppingcart.dao.CustomerDao;
 import woowacourse.shoppingcart.domain.Customer;
 import woowacourse.shoppingcart.dto.CustomerDto;
-import woowacourse.shoppingcart.dto.SignInDto;
+import woowacourse.auth.dto.SignInDto;
 import woowacourse.shoppingcart.dto.SignUpDto;
-import woowacourse.shoppingcart.dto.TokenResponseDto;
+import woowacourse.auth.dto.TokenResponseDto;
 import woowacourse.shoppingcart.dto.UpdateCustomerDto;
 import woowacourse.shoppingcart.exception.AuthorizationFailException;
 import woowacourse.shoppingcart.exception.DuplicateNameException;
